+++ lib/components/navigation_bar/components/m3e_nav_bar_destination_button.dart
@@ -209,6 +209,12 @@ class _M3ENavBarDestinationButtonState
               return null;
             },
           ),
+          ButtonActivateIntent: CallbackAction<ButtonActivateIntent>(
+            onInvoke: (ButtonActivateIntent intent) {
+              _select();
+              return null;
+            },
+          ),
         },
         child: Listener(
           behavior: HitTestBehavior.translucent,