+++ lib/components/navigation_drawer/components/m3e_drawer_destination_button.dart
@@ -136,6 +136,12 @@ class _M3EDrawerDestinationButtonState
                 return null;
               },
             ),
+            ButtonActivateIntent: CallbackAction<ButtonActivateIntent>(
+              onInvoke: (ButtonActivateIntent intent) {
+                _select();
+                return null;
+              },
+            ),
           },
           child: Listener(
             behavior: HitTestBehavior.translucent,